- Allow service to create hosts if they aren't found. This is configurable
