+++ README.md
@@ -23,7 +23,7 @@
 ## 5 Ws
 
 ### What
-DynCFDNS is a dynamic DNS update service for CloudFlare domains using Python. In short, it automatically updates your DNS records with your current external IP address at configurable intervals. It allows for multiple hostnames to be updated simultaneously, although all the records will point to the same external IP address. This is particularly useful for users with dynamic IP addresses who want to ensure their domains always resolve correctly. If a host record is not found, the applicatin can be configured to automatically create it in the specified domain, as long as you have the necessary permissions in your CloudFlare account.
+DynCFDNS is a dynamic DNS update service for CloudFlare domains using Python. In short, it automatically updates your DNS records with your current external IP address at configurable intervals. It allows for multiple hostnames to be updated simultaneously, although all the records will point to the same external IP address. This is particularly useful for users with dynamic IP addresses who want to ensure their domains always resolve correctly. If a host record is not found, the application can be configured to automatically create it in the specified domain, as long as you have the necessary permissions in your CloudFlare account.
 
 With DynCFDNS you no longer need external Dynamic DNS services if you own a domain. It directly integrates with CloudFlare's API to keep your DNS records up-to-date, ensuring your domains always point to the correct IP address. No more DynDNS, No-IP, or other third-party services needed.
 
@@ -38,7 +38,7 @@ I wanted to create a simple, efficient, and secure way to manage dynamic DNS upd
 DynCFDNS is developed and maintained by [AlxDroidDev](https://github.com/AlxDroidDev)
 
 ### Where
-São Paulo, SP, Brazil
+São Paulo - SP - Brazil
 
 ### When
 This project was started in July 2025.
@@ -72,13 +72,13 @@ This project was started in July 2025.
 ## Environment Variables
 
 | Variable               | Description                                                         | Required | Default | Example                               |
-|------------------------|---------------------------------------------------------------------|----------|---------|---------------------------------------|
-| `HOST_LIST`            | Comma-separated list of hostnames to update                         | ✅ | -       | `home.example.com,server.example.com` |
-| `CLOUDFLARE_API_TOKEN` | CloudFlare API Token                                                | ✅ | -       | `your_api_token_here`                 |
-| `CLOUDFLARE_API_KEY`   | CloudFlare Global API Key                                           | ✅ | -       | `your_api_key_here`                   |
-| `CLOUDFLARE_API_EMAIL` | CloudFlare account email                                            | ✅ | -       | `your-email@example.com`              |
-| `UPDATE_INTERVAL`      | Update interval in minutes                                          | ❌ | `60`    | `30`                                  |
-| `ALLOW_CREATE_HOSTS`   | Automatically create hosts in the given domain if they do not exist | ❌ | `false` | `true`                                |
+|------------------------|---------------------------------------------------------------------|:--------:|:-------:|---------------------------------------|
+| `HOST_LIST`            | Comma-separated list of hostnames to update                         |    ✔️    |    -    | `home.example.com,server.example.com` |
+| `CLOUDFLARE_API_TOKEN` | CloudFlare API Token                                                |    ✔️    |    -    | `your_api_token_here`                 |
+| `CLOUDFLARE_API_KEY`   | CloudFlare Global API Key                                           |    ✔️    |    -    | `your_api_key_here`                   |
+| `CLOUDFLARE_API_EMAIL` | CloudFlare account email                                            |    ✔️    |    -    | `your-email@example.com`              |
+| `UPDATE_INTERVAL`      | Update interval in minutes                                          |    ✖️    |  `60`   | `30`                                  |
+| `ALLOW_CREATE_HOSTS`   | Automatically create hosts in the given domain if they do not exist |    ✖️    | `false` | `true`                                |
 
 
 ### Getting CloudFlare Credentials
@@ -86,7 +86,7 @@ This project was started in July 2025.
 1. **API Token** (Required):
    - Go to [CloudFlare Dashboard](https://dash.cloudflare.com/profile/api-tokens)
    - Click "Create Token"
-   - Use tthe template "Edit zone DNS" or create a custom token with the following permissions:
+   - Use the template "Edit zone DNS" or create a custom token with the following permissions:
        - Permissions:
             - Zone: DNS -> Read
        - Zone Resources:
@@ -94,7 +94,7 @@ This project was started in July 2025.
        - Client IP Address filtering (Optional):
             - You can restrict the token to specific IP addresses for added security.
 
-![Cloudflare API token permissions](./images/cf_token_perms.png)
+      ![Cloudflare API token permissions](./images/cf_token_perms.png)
 
 2. **Global API Key**:
    - Go to [CloudFlare Dashboard](https://dash.cloudflare.com/profile/api-tokens)
@@ -137,21 +137,20 @@ docker run -d \
   -e HOST_LIST="home.example.com,server.example.com" \
   -e CLOUDFLARE_API_TOKEN="your_api_token" \
   -e CLOUDFLARE_API_KEY="your_api_key" \
-  -e CLOUDFLARE_API_EMAIL="your-email@example.com" \
-  -e UPDATE_INTERVAL="60" \
+  -e CLOUDFLARE_API_EMAIL="your-email@example.com" 
   dyncfdns
 ```
 
 ### Docker Compose
 
-Create a `docker-compose.yml` file:
+Create a minimal `docker-compose.yml` file (there are more configuration options available, see the [Full compose.yaml file](compose/compose.yml):
 
 ```yaml
 version: '3.8'
 
 services:
   dyncfdns:
-    build: .
+    image: alxdroiddev/dyncfdns:latest
     container_name: dyncfdns
     restart: unless-stopped
     environment:
@@ -159,10 +158,6 @@ services:
       - CLOUDFLARE_API_TOKEN=your_api_token
       - CLOUDFLARE_API_KEY=your_api_key
       - CLOUDFLARE_API_EMAIL=your-email@example.com
-      - UPDATE_INTERVAL=60
-    # Optional: Use environment file
-    # env_file:
-    #   - .env
 ```
 
 #### Using Environment File
@@ -175,20 +170,6 @@ CLOUDFLARE_API_TOKEN=your_api_token
 CLOUDFLARE_API_KEY=your_api_key
 CLOUDFLARE_API_EMAIL=your-email@example.com
 UPDATE_INTERVAL=120 # Optional. Value in seconds. Default is 120 seconds (2 minutes). This value also specifies the interval between healthchecks.
-```
-
-Update `docker-compose.yml`:
-
-```yaml
-version: '3.8'
-
-services:
-  dyncfdns:
-    build: .
-    container_name: dyncfdns
-    restart: unless-stopped
-    env_file:
-      - .env
 ```
 
 ### Run with Docker Compose
@@ -221,8 +202,9 @@ DynCFDNS/
 └── README.md            # This documentation
 ```
 
-## Quick Start
+## Quick Start 
 
+### By cloning the repository
 1. **Clone the repository**:
    ```bash
    git clone https://github.com/AlxDroidDev/DynCFDNS.git
@@ -244,6 +226,29 @@ DynCFDNS/
    ```bash
    docker-compose logs -f dyncfdns
    ```
+
+
+### By pulling Docker image
+
+1. **Pull the Docker image**:
+   ```bash
+   docker pull alxdroiddev/dyncfdns:latest
+   ```
+   
+2. **Run the container**:
+   ```bash
+    docker run -d \
+      --name dyncfdns \
+      --restart unless-stopped \
+      -e HOST_LIST="home.example.com,server.example.com" \
+      -e CLOUDFLARE_API_TOKEN="your_api" \
+      -e CLOUDFLARE_API_KEY="your_api_key" 
+      -e CLOUDFLARE_API_EMAIL="your cloudflare email" \
+      alxdroiddev/dyncfdns:latest
+   ```
+   OR
+
+   Use the docker compose file provided in the repository, and outlined above.
 
 ## Monitoring
